Clicking on question mark now show answer.
There's an event handler now for touch on the main screen that allows the user to touch on the question mark to see the answer.
Also, the controllers related to question registry were moved inside a group of them own. Question will go there when I create the object on core data.