Made tests pass when running from Xcode.
Test application was trying to get access to the kotoba database, which was making a test give a error for some reason (only when running on Xcode).
I changed the test to inject the data manager and the implementation class to detect if the data manager was already inject, and if it is to not start the data manager again.
Testing now seems to be working.