Bugfix: Application wouldn't start.
Application wouldn't start because of a BAD_ACCESS on tableView when the documentDidLoad called nextQuestion.

It was verified that every access to self caused a BAD_ACCESS. This was probably caused because the biding to self was lost when the framework dispatched a notification.

In order to fix this problem, intern was repaired to use a delegate to inform when the data manager did finished loading instead of NSNotification. Kotoba was then adapted to use this new architecture of intern.