Changed behavior of return key when user is creating a question.
When the user creates a new question or edit an existing question, there are two fields that the user must fill: the question field and the answer field.
There was a usability problem in the question field that the return key on the keyboard did nothing. This behavior was changed so that when the user touches on the return key, the focus will be changed to the answer field.
Also, the text in the return key was changed from "Return" to "Next".