Kotoba now starts in Portrait in both iPhone and IPad versions
The problem was in Info.plist file. The landscape orientation came before the portrait orientation. This way, the application always started on landscape and only after it was fully loaded it rotated to portrait.

As it is much more natural for users to hold the iPhone in portrait position, the application was changed to start in portrait, and if needed, rotate to landscape.