Automation Testing
Created a test to do automated acceptance testing of the question lifecycle. For now, it is still not running on jenkins, but I'll try to solve this later.

Also fixed a bug of position of stuff on the iPad. Now both text views of Question Game have scrolling (and it's working).